Modification of functions
